Move more tests under coveralls coverage - Get rid of the special-case test config environments: testTokenReq, testTokenNotReq, testTokenEnforced. - Use the “enforceApiToken” feature toggle instead of a config environment variable - Move the jobQueue and realtime under code coverage - Consolidate the tests which require USE_ACCESS_TOKEN=true into a single directory - Move the check for the “useWorkerProcess” feature toggle from the module level (only evaluated at startup) to the function level (evaluated at runtime) - Add feature toggle “enforceApiToken” to list of “longTermToggles” - Add “toggleOverride” function to testUtils so that tests can override a feature toggle at runtime - Update the jobQueue and realtime tests to use “toggleOverride” to set and unset the feature before and after running those tests
