+++ tests/enforceToken/createToken.js
@@ -12,9 +12,9 @@
 
 const expect = require('chai').expect;
 const supertest = require('supertest');
-const api = supertest(require('../../../../index').app);
-const constants = require('../../../../api/v1/constants');
-const u = require('../../../testUtils');
+const api = supertest(require('../../index').app);
+const constants = require('../../api/v1/constants');
+const u = require('../testUtils');
 const registerPath = '/v1/register';
 const tokenPath = '/v1/tokens';
