+++ tests/enforceToken/sampleUpsertBulk.js
