+++ tests/enforceToken/subjectGetHierarchy.js
